feat(audit): 添加操作日志管理功能
1. 新增操作日志菜单路由和侧边栏入口
2. 新增操作日志API接口和实体类定义
3. 新增操作日志列表页面，支持查询、分页和详情查看
4. 修复结算服务金额过滤参数类型不匹配问题

+++ his-rule-engine/his-settlement-service/src/main/java/com/his/settlement/service/SettlementService.java
@@ -174,8 +174,8 @@ public class SettlementService {
                 .eq(StringUtils.hasText(status), SettlementResult::getStatus, status)
                 .ge(startDate != null && !startDate.isBlank(), SettlementResult::getCreateTime, parseDate(startDate))
                 .le(endDate != null && !endDate.isBlank(), SettlementResult::getCreateTime, parseEndDate(endDate))
-                .ge(minAmount != null, SettlementResult::getTotalFee, BigDecimal.valueOf(minAmount))
-                .le(maxAmount != null, SettlementResult::getTotalFee, BigDecimal.valueOf(maxAmount))
+                .ge(minAmount != null, SettlementResult::getTotalFee, minAmount)
+                .le(maxAmount != null, SettlementResult::getTotalFee, maxAmount)
                 .orderByDesc(SettlementResult::getCreateTime);
 
         IPage<SettlementResult> pageResult = settlementMapper.selectPage(pageParam, wrapper);
